# Conflicts:
#	pay-channel/pay-alipay/src/main/java/org/dows/pay/alipay/AlipayPayHandler.java

+++ pay-biz/src/main/java/org/dows/pay/biz/payBiz.java
@@ -224,7 +224,7 @@ public class payBiz implements PayApi {
                 return PayQueryRes.builder()
                         .orderId(pay.getOrderId())
                         .outTradeNo(pay.getDealTo())
-                        .payAmount(pay.getAmount())
+                        .payAmount(pay.getAmount().intValue())
                         .payChannel(pay.getPayChannel())
                         .payDesc(pay.getTradeState())
                         .payTime(pay.getTransactionTime())
@@ -259,7 +259,7 @@ public class payBiz implements PayApi {
                     .outTradeNo(Optional.ofNullable(map.get("transaction_id")).map(Object::toString).orElse(null))
                     .payChannel(payTransaction.getPayChannel())
                     .orderId(payTransaction.getOrderId())
-                    .payAmount(payTransaction.getAmount())
+                    .payAmount(payTransaction.getAmount().intValue())
                     .build();
         }
         // todo:支付宝查询
